Remove `Network.Socket.ntohl` in favor of `System.ByteOrder.fromBigEndian`
`ntohl` has been removed from `network` as of v3, and although there's
some sense that it should be provided *somewhere* (and there are
issues and pull requests scattered around various places to this
effect), none of the concerned maintainers seem all that concerned
with doing anything.

So rather than wait for there to be unexpected motion on that front,
just side-step the whole issue by getting equivalent functionality
elsewhere.

A search for 'haskell endian' yields a number of packages that provide
this functionality; `byte-order` seemed small, focussed and had had a
recent update implying that it was actually alive.

This should resolve #4